@@ -1,29 +1,27 @@
 {
   "name": "egg-demo",
   "version": "1.0.0",
-  "description": "egg demo",
+  "description": "eggjs demo",
   "private": true,
   "egg": {
     "declarations": true
   },
   "dependencies": {
-    "egg": "^2.15.1",
-    "egg-scripts": "^2.11.0"
+    "egg": "^3",
+    "egg-scripts": "^2"
   },
   "devDependencies": {
-    "autod": "^3.0.1",
-    "autod-egg": "^1.1.0",
-    "egg-bin": "^4.11.0",
-    "egg-ci": "^1.11.0",
-    "egg-mock": "^3.21.0",
-    "eslint": "^5.13.0",
-    "eslint-config-egg": "^7.1.0"
+    "egg-bin": "^5",
+    "egg-ci": "^2",
+    "egg-mock": "^5",
+    "eslint": "^8",
+    "eslint-config-egg": "^12"
   },
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=16.0.0"
   },
   "scripts": {
-    "start": "egg-scripts start --daemon --title=egg-server-egg-demo --port 7005",
+    "start": "egg-scripts start --daemon --title=egg-server-egg-demo",
     "stop": "egg-scripts stop --title=egg-server-egg-demo",
     "dev": "egg-bin dev",
     "debug": "egg-bin debug",
@@ -31,16 +29,16 @@
     "test-local": "egg-bin test",
     "cov": "egg-bin cov",
     "lint": "eslint .",
-    "ci": "npm run lint && npm run cov",
-    "autod": "autod"
+    "ci": "npm run lint && npm run cov"
   },
   "ci": {
-    "version": "10"
+    "version": "16, 18",
+    "type": "github"
   },
   "repository": {
     "type": "git",
     "url": ""
   },
-  "author": "hisheng",
+  "author": "",
   "license": "MIT"
 }
