@@ -19,28 +19,27 @@
     "bookshelf": "^1.2.0",
     "chalk": "^4.1.2",
     "delegates": "^1.0.0",
-    "knex": "^0.95.9",
-    "koa": "^2.13.1",
-    "koa-body": "^4.2.0",
+    "knex": "^2.5.1",
+    "koa": "^2.14.2",
+    "koa-body": "^6.0.0",
     "koa-mount": "^4.0.0",
-    "koa-router": "^10.0.0",
-    "koa-session": "^6.2.0",
+    "koa-router": "^12.0.0",
+    "koa-session": "^6.4.0",
     "koa-static": "^5.0.0",
-    "koa-views": "^7.0.0",
-    "mysql2": "^2.3.0",
-    "node-schedule": "^2.0.0",
-    "nunjucks": "^3.2.3",
-    "sequelize": "^6.6.5"
+    "koa-views": "^8.1.0",
+    "mysql2": "^3.6.0",
+    "node-schedule": "^2.1.0",
+    "nunjucks": "^3.2.4",
+    "sequelize": "^6.33.0"
   },
   "devDependencies": {
-    "autocannon": "^7.4.0",
-    "ava": "^3.15.0",
-    "chai": "^4.3.4",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.24.0",
-    "mocha": "^9.1.4",
-    "supertest": "^6.1.4",
+    "autocannon": "^7.12.0",
+    "chai": "^4.3.8",
+    "eslint": "^8.49.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.28.0",
+    "mocha": "^10.2.0",
+    "supertest": "^6.3.3",
     "vitepress": "1.0.0-rc.14"
   },
   "repository": {
