@@ -1,14 +1,30 @@
+// ================= express =================
+
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
+import * as config from '../../config.js';
+
+async function bootstrap() {
+  const app = await NestFactory.create(AppModule);
+  await app.listen(config.nest_fastify);
+}
+bootstrap();
+
+// ================= fastify =================
+
+/* import { NestFactory } from '@nestjs/core';
 import {
   FastifyAdapter,
   NestFastifyApplication,
 } from '@nestjs/platform-fastify';
-
-const config = require('../../config');
+import { AppModule } from './app.module';
+import * as config from '../../config.js';
 
 async function bootstrap() {
-  const app = await NestFactory.create<NestFastifyApplication>(AppModule, new FastifyAdapter());
-  await app.listen(config.nest_fastify.port);
+  const app = await NestFactory.create<NestFastifyApplication>(
+    AppModule,
+    new FastifyAdapter(),
+  );
+  await app.listen(config.nest_fastify);
 }
-bootstrap();
+bootstrap(); */
