@@ -3,7 +3,7 @@
 const { app, assert } = require('egg-mock/bootstrap');
 
 describe('test/app/controller/home.test.js', () => {
-  it('should assert', () => {
+  it('should assert', async () => {
     const pkg = require('../../../package.json');
     assert(app.config.keys.startsWith(pkg.name));
 
@@ -11,7 +11,7 @@ describe('test/app/controller/home.test.js', () => {
     // yield ctx.service.xx();
   });
 
-  it('should GET /', () => {
+  it('should GET /', async () => {
     return app.httpRequest()
       .get('/')
       .expect('hi, egg')
