@@ -3,5 +3,5 @@ const config = require('../config');
 module.exports = {
   debugFlag: 'Hisheng',
   port: config.heysoo.port,
-  logRequestTime: false,
+  logRequestTime: true,
 };
