@@ -1,7 +1,7 @@
 // default config of Application
 
 module.exports = {
-  host: '127.0.0.1',
+  host: 'localhost',
   port: 8686,
   folder: {
     app: 'app',
