@@ -1,7 +1,7 @@
 const Koa = require('koa');
 const path = require('path');
 const assert = require('assert');
-const koaBody = require('koa-body');
+const { koaBody } = require('koa-body');
 
 // extends for Koa
 const request = require('./extend/request.js');
